requirements.txt
streamlit
transformers
torch
Pillow
huggingface_hub
